[jit] Better match behavior of loaded ScriptModules vs. freshly created ones
IR emitter uses `ModuleValue` to represent ScriptModules and emit IR for
attribute access, submodule access, etc.

`ModuleValue` relies on two pieces of information, the JIT type of the
module, and the `ConcreteModuleType`, which encapsulates Python-only
information about the module.

ScriptModules loaded from a package used to create a dummy
ConcreteModuleType without any info in it. This led to divergences in
behavior during compilation.

This PR makes the two ways of constructing a ConcreteModuleType equivalent,
modulo any py-only information (which, by definition, is never present in
packaged files anyway).

ghstack-source-id: e0c6e8290e343edd71c3119ccb84f77097cd122d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/43298